requirements.txt
flask
pickle
requests
python-telegram-bot